Append more info to the code

+++ OperacoesMatematicas.java
@@ -15,7 +15,7 @@ public class OperacoesMatematicas {
 		return num1 * num2;
 	}
 	
-	public float dividir(int num1, int num2)
+	public int dividir(int num1, int num2)
 	{
 		return num1 / num2;
 	}
